+++ libnymea-core/integrations/pythonintegrationplugin.h
@@ -5,6 +5,7 @@
 
 #include <QObject>
 #include <QJsonObject>
+#include <QFuture>
 
 extern "C" {
 typedef struct _object PyObject;
@@ -17,6 +18,7 @@ class PythonIntegrationPlugin : public IntegrationPlugin
     Q_OBJECT
 public:
     explicit PythonIntegrationPlugin(QObject *parent = nullptr);
+    ~PythonIntegrationPlugin();
 
     static void initPython();
 
@@ -41,6 +43,7 @@ private:
     QVariantMap m_metaData;
     PyObject *m_module = nullptr;
     PyThreadState *m_interpreter = nullptr;
+    QFuture<void> m_eventLoop;
 
 };
