+++ libnymea-core/libnymea-core.pro
@@ -21,6 +21,7 @@ RESOURCES += $$top_srcdir/icons.qrc \
 HEADERS += nymeacore.h \
     integrations/plugininfocache.h \
     integrations/python/pything.h \
+    integrations/python/pythingdescriptor.h \
     integrations/python/pythingdiscoveryinfo.h \
     integrations/python/pythingsetupinfo.h \
     integrations/thingmanagerimplementation.h \
