+++ libnymea/loggingcategories.h
@@ -89,6 +89,9 @@ Q_DECLARE_LOGGING_CATEGORY(dcMqtt)
 Q_DECLARE_LOGGING_CATEGORY(dcTranslations)
 Q_DECLARE_LOGGING_CATEGORY(dcCoap)
 Q_DECLARE_LOGGING_CATEGORY(dcI2C)
+Q_DECLARE_LOGGING_CATEGORY(dcIntegrations)
+Q_DECLARE_LOGGING_CATEGORY(dcJsIntegrations)
+Q_DECLARE_LOGGING_CATEGORY(dcPythonIntegrations)
 
 
 /*
